% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.zph.R
\name{cox.zph}
\alias{cox.zph}
\title{Test the proportional hazards assumption of an RPSFTM/Cox Regression}
\usage{
cox.zph(fit, ...)
}
\arguments{
\item{fit}{the result of fitting a rpsftm model using \code{\link[survival]{coxph}} as the inner estimation tool.}

\item{...}{any other arguments to pass to \code{\link[survival]{cox.zph}}.}
}
\description{
If the the \code{fit} inherits *both* \code{rpsftm} and \code{coxph} then this pulls out the genuine 
\code{survival::coxph} object that is deeply nested in the object, and then runs \code{survival::cox.zph} on it.
Or it avoids overwriting the function from \code{survival} by calling \code{survival::cox.zph} directly if the object 
does not inherit \code{rpsftm}. Or it fails.
}
\note{
This does rely on the order of loading packages. The \code{rpsftm} package must be loaded after \code{survival}, 
if both are required, to avoid the masking of synonymous functions causing errors.
}
\seealso{
\code{\link[survival]{cox.zph}}
}
