% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crpsftm.R
\name{crpsftm}
\alias{crpsftm}
\title{Connie's Rank Preserving Structural Failure Time Model}
\usage{
crpsftm(data, si)
}
\arguments{
\item{data}{Data in her format.}

\item{si}{Grid of potential parameter values to estimate over. By default, the grid is set to \code{seq(-3,3,.001)}.}
}
\value{
From the potential values of psi, the function will return a list with her preferred items.
}
\description{
\code{crpsftm} implements the rank preserving structural failure time model presented by Robins and Tsiatis (1991) and again by Korhonen et al (2012).
}
\details{
Data should be provided in her format.
}
\author{
Linh Tran
}

