% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_eqn.R
\name{est_eqn}
\alias{est_eqn}
\title{Estimating Equations for \code{rpsftm()}}
\usage{
est_eqn(psi, data, formula, target = 0, test = "survdiff", recensor,
  autoswitch, ...)
}
\arguments{
\item{psi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{data}{the data set that contains the variables. 
Must contain columns named: time, censor_time,rx, arm.
Optionally a column named: treat_modifier}

\item{formula}{a forumula object of covariates to adjust for: \code{~strata(A)+B*C}}

\item{target}{the value to subtract off from the z-statistic}

\item{test}{the survival regression function to calculate the z-statistic: survdiff, coxph, survreg}

\item{recensor}{a logical to use recensoring if set to TRUE. Default is TRUE. If FALSE then
the status of censoring or event is unchanged, but the time and censor_time values are both transformed according
to the proportion of time on treatment. If FALSE then recensoring is applied at the earliest of either 
a transformed or untransformed censor_time. recensoring avoids the bias induced by 
informative censoring on the transformed scale.}

\item{autoswitch}{a logical to autodetect cases of no switching. Default is TRUE. If all observations in an arm
have perfect compliance then recensoring is not applied in that arm. If FALSE the recensoring is applied
regardless of perfect compliance.}

\item{...}{arguments to supply to the test function.}
}
\value{
A scalar value of the estimating equation: the z-statistics from a test minus a target value
}
\description{
Calculates the Estimating Equation to be solved in RPSFTM models
}
\author{
Simon Bond
}
\seealso{
\code{\link{recensor}}
}

