% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EstEqn.R
\name{EstEqn}
\alias{EstEqn}
\title{Estimating Equations for \code{rpsftm()}}
\usage{
EstEqn(phi, time, censor_time, rx, data, armName, formula, target = 0,
  test = "survdiff", ...)
}
\arguments{
\item{phi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{time}{the observed failure or censoring time.}

\item{censor_time}{the theoretical censoring time, either observed or set after time.}

\item{rx}{the proportion of time spent on treatment}

\item{data}{a data.frame object containing the variables}

\item{target}{the value to subtract off from the z-statistic}

\item{test}{the survival regression function to calculate the z-statistic: survdiff, coxph, survreg}

\item{arm}{the randomisation that is independent of the recensored survival times}

\item{adjustors}{a forumula object of covariates to adjust for: \code{~strata(A)+B*C}}

\item{\code{...}}{arguments to supply to the test function.}
}
\value{
A scalar value of the estimating equation: the z-statistics from a test minus a target value
}
\description{
Calculates the Estimating Equation to be solved in RPSFTM models
}
\author{
Simon Bond
}
\seealso{
recensor
}

