% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chisq.R
\name{extract_chisq}
\alias{extract_chisq}
\title{Extracting the chi-squared statistic from a survival object}
\usage{
extract_chisq(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{extendible arguments to the S3 method}
}
\value{
a numeric value, the chi-squared statistic for the independence test of the treatment arms
}
\description{
Generic S3 methods to extract the z-statistic from a set of survival fit objects
}
\author{
Simon Bond
}
\keyword{internal}
