% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generating simulated clinical trial data}
\usage{
generateData(n, lambda, time, p = 0.5, alpha, txEffect = "accel", c = 0,
  crossover = 0, long = FALSE)
}
\arguments{
\item{n}{The number of observations desired to generate}

\item{lambda}{The set of hazards to be used in generating the data}

\item{time}{The number of time points to generate the data over}

\item{p}{The proportion of patients being assigned active treatment.}

\item{alpha}{The effect of the treatment, under the model specified. Under the \code{ldr} assumption, two values will have to be supplied: (1) an effect of treatment, and (2) a parity time point at which there is no longer a treatment effect.}

\item{txEffect}{The statistical model assumed for the treatment effect to be applied. \code{accel} will accelerate/decelerate the survival curve by \code{exp(alpha)}. \code{propor} will increase/decrease the hazard rates by a factor of \code{exp(alpha)}. \code{additive} will increase/decrease the hazard rates by adding \code{alpha}. \code{ldr} will increase/decrease the hazards by a factor of \code{exp(alpha)} which decreases over time until a parity time is met (at which point there is no treatment effect and hazard rates between the two treatment groups are equal). If the \code{ldr} treatment is specified,}

\item{c}{proportion of subjects to be censored at each time point. Currently assumes censoring to be independent of survival time.}

\item{long}{logical; If \code{TRUE}, the data set returned will be in long format (as opposed to wide).}
}
\value{
An object of class \code{simData} will be returned, which includes both the data and the hazards used to generate the observations.
}
\description{
\code{generateData} generates observations which mimick data observed in a clinical trial.
}
\details{
This code will generate data based on an underlying baseline set of hazards, an effect value, and an assumed statistical model. In being fed the set of hazards, the code assumes that survival will follow a piecewise constant survival distribution (i.e. a piecewise exponential distribution). Possible treatment effects include an accelerated factor, proportional hazards, additive hazards, and log-linear declining relative hazard.
}
\author{
Linh Tran
}

