% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specials.R
\name{Instr}
\alias{Instr}
\title{Special Functions to use in the rpsftm() formula}
\usage{
Instr(arm, rx)
}
\arguments{
\item{arm}{the randomised treatment arm. a factor with 2 levels, or numeric variable with values 0/1.}

\item{rx}{the proportion of time on active treatment (arm=1 or the non-reference level of the factor)}
}
\value{
matrix with two columns named either time and censor_time, or, arm and rx. These can be used in the formula argument to rpsftm()
}
\description{
A couple of functions that are defined to be used in the formula argument, 
and identifed as specials in the terms() object
}
\section{Functions}{
\itemize{
\item \code{Instr}: special function
}}
\author{
Simon Bond
}

