% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipe.R
\name{ipe}
\alias{ipe}
\title{Iterative Parameter Estimate method}
\usage{
ipe(
  formula,
  data,
  censor_time,
  subset,
  na.action,
  dist = "weibull",
  treat_modifier = 1,
  autoswitch = TRUE,
  epsilon = 1e-05,
  iter_max = 100,
  alpha = 0.05,
  verbose = options()$verbose,
  ...
)
}
\arguments{
\item{formula}{a formula with a minimal structure of \code{Surv(time, status)~rand(arm,rx)}.
Further terms can be added to the right hand side to adjust for covariates and use strata or 
cluster arguments.}

\item{data}{an optional data frame that contains variables}

\item{censor_time}{variable or constant giving the time at which censoring would, or has occurred. 
This should be provided for all observations unlike standard Kaplan-Meier or Cox regression where 
it is only given for censored observations. If no value is given then re-censoring is not applied.}

\item{subset}{expression indicating which subset of the rows of data should be used in the fit. 
This can be a logical vector (which is replicated to have length equal to the number of observations),
a numeric vector indicating which observation numbers are to be included (or excluded if negative), 
or a character vector of row names to be included. All observations are included by default.}

\item{na.action}{a missing-data filter function. This is applied to the model.frame after any subset 
argument has been used. Default is options()$na.action.}

\item{dist}{
assumed distribution for y variable. 
If the argument is a character string, then it is assumed to name an
element from \code{\link[survival]{survreg.distributions}}. These include
\code{"weibull"}, \code{"exponential"}, \code{"gaussian"},
\code{"logistic"},\code{"lognormal"} and \code{"loglogistic"}.
Otherwise, it is assumed to be a user defined list conforming to the
format described in \code{\link[survival]{survreg.distributions}}.
}

\item{treat_modifier}{an optional variable that psi is multiplied by on an individual observation level to give
differing impact to treatment.}

\item{epsilon}{the threshold to define convergence of the parameter estimate psi.}

\item{iter_max}{the maximum number of iterations allowed before termination.}

\item{alpha}{the significance level used to calculate confidence intervals}

\item{...}{other arguments that will be passed to \code{survreg}}
}
\value{
a ipe method object that is a list of the following:
\itemize{
\item psi: the estimated parameter
\item fit: a survdiff object to produce Kaplan-Meier curves of the estimated counterfactual failure times for each treatment arm assuming perfect compliance
\item Sstar: the recensored \code{Surv()} data using the estimate value of psi to give counterfactual failure times with perfect compliance.
\item rand: the rand() object used to specify the allocated and observed amount of treatment.
\item iter: the number of iterations taken to achieve convergence
\item Further elements corresponding to either a \code{survreg} object. This will always include:
  \itemize{
  \item call: the R call object
  \item formula: a formula representing any adjustments, strata or clusters- used for the \code{update()} function
  \item terms: a more detailed representation of the model formula
  }
}
}
\description{
Main Function used for estimating causal parameters under the Iterative Parameter Estimate method
}
\details{
the formula object \code{Surv(time, status)~rand(arm,rx)}. \code{rand()} stands 
for randomisation, both the randomly assigned and actual observed treatment. 
\itemize{
  \item arm: the randomised treatment arm. a factor with 2 levels, or numeric variable with values 0/1.
 \item rx: the proportion of time on active treatment (arm=1 or the non-reference level of the factor)
}
Further adjustment terms can be added on the right hand side of the formula if desired, included \code{strata()}
or \code{cluster()} terms.
}
\examples{
?immdef
fit <- ipe(Surv(progyrs, prog)~rand(imm,1-xoyrs/progyrs),immdef, censyrs)
print(fit)
summary(fit)
plot(fit)

}
\seealso{
\code{\link[survival]{survdiff}}, \code{\link[survival]{coxph.object}}, \code{\link[survival]{survreg.object}}
}
\author{
Simon Bond
}
