% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{old_rand}
\alias{old_rand}
\title{rand functions to use in the rpsftm() formula}
\usage{
old_rand(arm, rx)
}
\arguments{
\item{arm}{the randomised treatment arm. a factor with 2 levels, or numeric variable with values 0/1.}

\item{rx}{the proportion of time on active treatment (arm=1 or the non-reference level of the factor)}
}
\value{
matrix with two columns named arm and rx. These can be used in the formula argument to rpsftm()
}
\description{
A function that is defined to be used in the formula argument, 
and identified as specials in the terms() object
}
\seealso{
\code{\link{print.rand}}, \code{\link{rpsftm}}
}
\author{
Simon Bond
}
