% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rpsftm.R
\name{print.rand}
\alias{print.rand}
\title{Print method}
\usage{
\method{print}{rand}(x, ...)
}
\arguments{
\item{x}{a rand() object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of rx values broken down by arm for a rand() object
}
\description{
print method for rand() objects - to display the summary of rx, by arm
}
\seealso{
\code{\link{rand}}, \code{\link{rpsftm}}
}
\author{
Simon Bond
}
