% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rpsftm.R
\name{print.rpsftm.coxph}
\alias{print.rpsftm.coxph}
\title{modified version of print.coxph}
\usage{
\method{print}{rpsftm.coxph}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
\item{x}{an coxph()}

\item{digits}{the number of digits to print out}

\item{...}{further arguments}
}
\description{
this drops the "arm" term as this is not a real parameter
}
\keyword{internal}
