% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.rpsftm.R
\name{print.rpsftm}
\alias{print.rpsftm}
\title{Print Method}
\usage{
\method{print}{rpsftm}(x)
}
\arguments{
\item{x}{an object returned from the \code{rpsftm()} function}
}
\value{
a print of the underlying test object at the point estimate
}
\description{
Function used to plot the KM curves of the treatment-free transformed times
}
\author{
Simon Bond
}

