% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rpsftm.R
\name{print.rpsftm.summary.coxph}
\alias{print.rpsftm.summary.coxph}
\title{modified version of print.summary.coxph}
\usage{
\method{print}{rpsftm.summary.coxph}(x, digits = max(getOption("digits")
  - 3, 3), signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{the result of a call to \code{summary.coxph}}

\item{digits}{significant digits to print}

\item{signif.stars}{show star to highlight small p-values}

\item{...}{further arguments for future methods}
}
\description{
this drops the "arm" term as this is not a real parameter
}
\keyword{internal}
