% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rpsftm.R
\name{print.summary.survreg}
\alias{print.summary.survreg}
\title{modified version of print.summary.survreg}
\usage{
\method{print}{summary.survreg}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
\item{x}{the result of a call to \code{summary.survreg}}

\item{digits}{significant digits to print}

\item{...}{further arguments for future methods}
}
\description{
this drops the "arm" term as this is not a real parameter
}
\keyword{internal}

