% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rpsftm.R
\name{print.survreg}
\alias{print.survreg}
\title{modified version of print.survreg}
\usage{
\method{print}{survreg}(x, ...)
}
\arguments{
\item{x}{a survreg() object}

\item{...}{further arguments}
}
\description{
this drops the "arm" term as this is not a real parameter
}
\keyword{internal}

