% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rpsftm.R
\name{print_rand}
\alias{print_rand}
\title{Print method}
\usage{
print_rand(x)
}
\arguments{
\item{x}{a rpsftm object}
}
\value{
a summary of treatment received values broken down by arm.
}
\description{
print method for rand() objects - to display the summary of rx, by arm
}
\seealso{
\code{\link{rpsftm}}
}
\author{
Simon Bond
}
