% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recensor.R
\name{recensor}
\alias{recensor}
\title{Recensoring}
\usage{
recensor(phi, time, censor_time, rx)
}
\arguments{
\item{phi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{time}{the observed failure or censoring time.}

\item{censor_time}{the theoretical censoring time, either observed or set after time.}

\item{rx}{the proportion of time spent on treatment}

\item{data}{a data.frame object containing the variables}
}
\value{
A Surv() object with the recensoring applied
}
\description{
Applies the Recensoring method using a given parameter value and returns a Surv() object
}
\author{
Simon Bond
}

