% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.rpsftm.R
\name{residuals.rpsftm}
\alias{residuals.rpsftm}
\title{residual() method for rpsftm objects}
\usage{
\method{residuals}{rpsftm}(object, ...)
}
\arguments{
\item{object}{an object returned from the \code{rpsftm()} function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a residuals object.
}
\description{
Function to apply residual method to rpsftm objects
}
\seealso{
\code{\link[stats]{residuals}} \code{\link[survival]{residuals.coxph}}  \code{\link[survival]{residuals.survreg}}
}
\author{
Simon Bond
}
