% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsftm-package.R
\docType{package}
\name{rpsftm-package}
\alias{rpsftm-package}
\title{rpsftm: a package to fit Rank Preserving Structural Failure Time Model}
\description{
This implements the method of Robins JM, Tsiatis AA. 
The key function is \code{\link{rpsftm}}, which provides estimates of the causal parameter of interest.
}
\details{
rpfstm: a package to fit Rank Preserving Structural Failure Time Model
}
\references{
Robins JM, Tsiatis AA. Correcting for non-compliance in randomized trials 
using rank preserving structural failure time models. 
Communications in Statistics--Theory and Methods 1991; 20: 2609--2631
}
\seealso{
\code{\link[survival]{survdiff}}

\code{\link[survival]{coxph}}

\code{\link[survival]{survreg}}
}
