% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rpsftm.R
\name{summary.rpsftm}
\alias{summary.rpsftm}
\title{summary Method}
\usage{
\method{summary}{rpsftm}(object, ...)
}
\arguments{
\item{object}{an object returned from the \code{rpsftm()} function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted regression model.
}
\description{
Function used to summarise the fitted model to an rpsftm object
}
\author{
Simon Bond
}

