% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.rpsftm.R
\name{survfit.rpsftm}
\alias{survfit.rpsftm}
\title{survfit() method for rpsftm objects}
\usage{
\method{survfit}{rpsftm}(formula, ...)
}
\arguments{
\item{formula}{an object returned from the \code{rpsftm()} function. (with the name "formula" to agree with generic argument names)}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a survfit object.
}
\description{
Function to apply survfit method to rpsftm objects
}
\seealso{
\code{\link[survival]{survfit}}
}
\author{
Simon Bond
}
