% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.inner.R
\name{terms.inner}
\alias{terms.inner}
\title{terms.inner}
\usage{
\method{terms}{inner}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{extendible arguments to the S3 method}
}
\value{
a list of variables referred to in a formula or call object

a character vector
}
\description{
A function taken from the survival library - it is not exported from there hence a local copy
}
\keyword{internal}
