% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_eqn.R, R/min_eqn.R
\name{est_eqn}
\alias{est_eqn}
\alias{min_eqn}
\title{Estimating Equations for \code{rpsftm()}}
\usage{
est_eqn(
  psi,
  response,
  data,
  formula_list,
  treatment_matrix,
  rand_matrix,
  target = 0,
  test = "survdiff",
  autoswitch,
  ...
)

min_eqn(
  psi,
  response,
  data,
  formula_list,
  treatment_matrix,
  rand_matrix,
  target = 0,
  test = "survdiff",
  autoswitch,
  ...
)
}
\arguments{
\item{psi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{response}{the response variable, a Surv() object}

\item{data}{the data set that contains the variables. 
Must contain columns named: time, censor_time,rx, arm.
Optionally a column named: treat_modifier}

\item{treatment_matrix}{a model.matrix object capturing the amount of active treatment(s) received}

\item{rand_matrix}{a model.matrix object catpuring the randomised treatment arms. Unused at presnet}

\item{target}{the value to subtract off from the z-statistic}

\item{test}{the survival regression function to calculate the z-statistic: survdiff, coxph, survreg}

\item{autoswitch}{Currently has no effect ( formerly: a logical to autodetect cases of no switching. If TRUE, then if all observations in an arm
have perfect compliance then recensoring is not applied in that arm. If FALSE the recensoring is applied
regardless of perfect compliance.)}

\item{...}{arguments to supply to the test function.}

\item{formula}{a formula object of covariates to adjust for: \code{~strata(A)+B*C}}

\item{rand}{a one-side formula giving the randomised arms}

\item{treatment}{a one-sided formula giving the amount of treatment recieved. 
The model matrix it produces must be of the same rank as for the \code{rand} argument. So either remove the intercept with '-1',
 or only provide \code{(k-1)} terms on a scale \emph{relative} to the control treatment, where \code{k} is the number of arms.}
}
\value{
A scalar value of the estimating equation: the z-statistics from a test minus a target value

A scalar value of the estimating equation: the z-statistics from a test minus a target value
}
\description{
Calculates the Estimating Equation to be solved in RPSFTM models

Calculates the Estimating Equation to be solved in RPSFTM models
}
\seealso{
\code{\link{untreated}}

\code{\link{untreated}}
}
\author{
Simon Bond

Simon Bond
}
\keyword{internal}
