% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EstEqn.R
\name{EstEqn}
\alias{EstEqn}
\title{Estimating Equations for \code{rpfsftm()}}
\usage{
EstEqn(phi, Time, CensorTime, Rx, data, Arm, adjustors = NULL, target = 0,
  test = survdiff, ...)
}
\arguments{
\item{phi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{Time}{the observed failure or censoring time.}

\item{CensorTime}{the theoretical censoring time, either observed or set after Time.}

\item{Rx}{the proportion of Time spent on treatment}

\item{data}{a data.frame object containing the variables}

\item{Arm}{the randomisation that is independent of the recensored survival times}

\item{adjustors}{a forumula object of covariates to adjust for: \code{~strata(A)+B*C}}

\item{target}{the value to subtract off from the z-statistic}

\item{test}{the survival regression function to calculate the z-statistic: survdiff, coxph, survreg}

\item{\code{...}}{arguments to supply to the test function.}
}
\value{
A scalar value of the estimating equation
}
\description{
Calculates the Estimating Equation to be solved in RPSFTM models
}
\author{
Simon Bond
}

