% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstEqn.R
\name{EstEqn}
\alias{EstEqn}
\title{Estimating Equations for \code{rpsftm()}}
\usage{
EstEqn(phi, time, censor_time, rx, data, arm, formula, target = 0,
  test = "survdiff", Recensor, Autoswitch, ...)
}
\arguments{
\item{phi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{time}{the observed failure or censoring time.}

\item{censor_time}{the theoretical censoring time, either observed or set after time.}

\item{rx}{the proportion of time spent on treatment}

\item{data}{the data set that contains the variables.}

\item{arm}{the randomised arm. Either a numerical indicator, with 0 as the placebo, or a factor with the lowest level as placebo.}

\item{formula}{a forumula object of covariates to adjust for: \code{~strata(A)+B*C}}

\item{target}{the value to subtract off from the z-statistic}

\item{test}{the survival regression function to calculate the z-statistic: survdiff, coxph, survreg}

\item{Recensor}{a logical to use recensoring if set to TRUE. Default is TRUE.}

\item{Autoswitch}{a logical to autodetect cases of no switching. Default is TRUE}

\item{\code{...}}{arguments to supply to the test function.}
}
\value{
A scalar value of the estimating equation: the z-statistics from a test minus a target value
}
\description{
Calculates the Estimating Equation to be solved in RPSFTM models
}
\author{
Simon Bond
}
\seealso{
recensor
}

