% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specials.R
\name{Instr}
\alias{Instr}
\alias{ReCen}
\title{Special Functions to use in the rpsftm() formula}
\usage{
Instr(arm, rx)

ReCen(time, censor_time)
}
\arguments{
\item{arm}{the randomised treatment arm. a factor with 2 levels, or numeric variable with values 0/1.}

\item{rx}{the proportion of time on active treatment (arm=1 or the non-reference level of the factor)}

\item{time}{the observed failure or censoring time}

\item{censor_time}{the time at which censoring would, or has occurred. This is provided for all observations
unlike standard Kaplan-Meier or Cox regression where it is only given for censored observations}
}
\value{
matrix with two columns named either time and censor_time, or, arm and rx. These can be used in the formula argument to rpsftm()
}
\description{
A couple of functions that are defined to be used in the formula argument, 
and identifed as specials in the terms() object
}
\section{Functions}{
\itemize{
\item \code{Instr}: Instr function

\item \code{ReCen}: ReCen function
}}
\author{
Simon Bond
}

