% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rpsftm.R
\name{plot.rpsftm}
\alias{plot.rpsftm}
\title{Plot Method}
\usage{
\method{plot}{rpsftm}(x, ...)
}
\arguments{
\item{x}{an object returned from the \code{\link{rpsftm}} function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a ggplot plot of the fitted KM curves.
}
\description{
Function used to plot the KM curves of the treatment-free transformed times
}
\author{
Simon Bond
}
