% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rpsftm.R
\name{print.rpsftm}
\alias{print.rpsftm}
\alias{print.ipe}
\title{Print Method}
\usage{
\method{print}{ipe}(x, ...)
}
\arguments{
\item{x}{an object returned from the \code{\link{rpsftm}} function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a print of the underlying test object at the point estimate.
}
\description{
Function used to print of the underlying test object at the point estimate of a rpsftm object
}
\author{
Simon Bond
}
