% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rpsftm.R
\name{print.rpsftm}
\alias{print.rpsftm}
\title{Print Method}
\usage{
\method{print}{rpsftm}(x)
}
\arguments{
\item{x}{an object returned from the \code{\link{rpsftm}} function}
}
\value{
a print of the underlying test object at the point estimate
}
\description{
Function used to print of the underlying test object at the point estimate of a rpsftm object
}
\author{
Simon Bond
}

