% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Recensor.R
\name{Recensor}
\alias{Recensor}
\title{Recensoring}
\usage{
Recensor(phi, Time, CensorTime, Rx, data)
}
\arguments{
\item{phi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{Time}{the observed failure or censoring time.}

\item{CensorTime}{the theoretical censoring time, either observed or set after Time.}

\item{Rx}{the proportion of Time spent on treatment}

\item{data}{a data.frame object containing the variables}
}
\value{
A Surv() object with the recensoring applied
}
\description{
Applies the Recensoring method using a given parameter value and returns a Surv() object
}
\author{
Simon Bond
}

