% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recensor.R
\name{recensor}
\alias{recensor}
\title{Recensoring}
\usage{
recensor(phi, time, censor_time, rx, arm, Recensor, Autoswitch)
}
\arguments{
\item{phi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{time}{the observed failure or censoring time.}

\item{censor_time}{the theoretical censoring time, either observed or set after time.}

\item{rx}{the proportion of time spent on treatment}

\item{arm}{the randomised arm. Either a numerical indicator, with 0 as the placebo, or a factor with the lowest level as placebo.}

\item{Recensor}{a logical to use recensoring if set to TRUE. Default is TRUE.}

\item{Autoswitch}{a logical to autodetect cases of no switching. Default is TRUE}
}
\value{
A Surv() object with the recensoring applied
}
\description{
Applies the Recensoring method using a given parameter value and returns a Surv() object
}
\author{
Simon Bond
}

