\name{rpsftm-package}
\alias{rpsftm-package}
\alias{rpsftm}
\docType{package}
\title{
Rank Preserving Structural Failure Time Models
}
\description{
This implements the method of Robins JM, Tsiatis AA. Correcting for non-compliance in randomized trials using rank
preserving structural failure time models. Communications in Statistics--Theory and
Methods 1991; 20: 2609--2631

}
\details{
\tabular{ll}{
Package: \tab rpsftm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-15\cr
License: \tab GPL-2\cr
}
The key function is \code{rpsftm()}, which provides estimates of the causal parameter of interest.
}
\author{
Simon Bond

Maintainer: Simon Bond <simon.bond@addenbrookes.nhs.uk>

}
\references{
Robins JM, Tsiatis AA. Correcting for non-compliance in randomized trials using rank
preserving structural failure time models. Communications in Statistics--Theory and
Methods 1991; 20: 2609--2631
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package 
models
survival
}
\seealso{
\code{\link[<pkg>:<pkg>-package]{survival}}
}
\examples{
~~ simple examples of the most important functions ~~
}
