% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsftm-package.R
\docType{package}
\name{rpsftm-package}
\alias{rpsftm-package}
\title{rpsftm: Rank Preserving Structural Failure Time Models}
\description{
Implements methods described by the paper Robins and Tsiatis (1991) \doi{10.1080/03610929108830654}. These use g-estimation to estimate the causal effect of a treatment in a two-armed randomised control trial where non-compliance exists and is measured, under an assumption of an accelerated failure time model and no unmeasured confounders.
}
\author{
\strong{Maintainer}: Simon Bond \email{simon.bond7@nhs.net} (primary author of code, secondary author of vignette)

Authors:
\itemize{
  \item Annabel Allison (primary author of vignette, secondary author of code)
}

}
\keyword{internal}
