% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsftm.R
\name{rpsftm}
\alias{rpsftm}
\title{Rank Preserving Structural Failure Time Model}
\usage{
rpsftm(formula, data, subset, na.action, test = survdiff, lowpsi = -1,
  hipsi = 1, alpha = 0.05, treat_weight = 1, Recensor = TRUE,
  Autoswitch = TRUE, ...)
}
\arguments{
\item{formula}{a formula with a minimal structure of \code{ReCen(time, censor_time)~Instr(arm,rx)}.
Further terms can be added to the right hand side to adjust for covariates and use strata or cluster arguments.}

\item{data}{an optional data frame that contains variables}

\item{subset}{expression indicating which subset of the rows of data should be used in the fit. 
This can be a logical vector (which is replicated to have length equal to the number of observations)
, a numeric vector indicating which observation numbers are to be included (or excluded if negative), 
or a character vector of row names to be included. All observations are included by default.}

\item{na.action}{a missing-data filter function. This is applied to the model.frame after any subset 
argument has been used. Default is options()$na.action.}

\item{test}{the survival regression function to calculate the z-statistic: survdiff, coxph, survreg}

\item{lowpsi}{the lower limit of the range to search for the causal parameter}

\item{hipsi}{the upper limit of the range to search for the causal paramater}

\item{alpha}{the significance level used to calculate confidence intervals}

\item{treat_weight}{an optional parameter that psi is multiplied by on an individual observation level to give
differing impact to treatment. The values are transformed by \code{abs(.)/max(abs(.))} to ensure 1 is the largest weight.}

\item{Recensor}{a logical to use recensoring if set to TRUE. Default is TRUE.}

\item{Autoswitch}{a logical to autodetect cases of no switching. Default is TRUE}

\item{...}{arguments to supply to the test function.}
}
\value{
a rpsftm method object that is a list of the following:
\itemize{
\item psi: the estimated parameter
\item fit: a survdiff object to produce Kaplan-Meier curves of the estimated counterfactual untreated failure times for each treatment arm
\item formula: a formula representing any adjustments, strata or clusters- used for the \code{update()} function
\item regression: the survival regression object at the estimated value of psi
\item Sstar: the recensored \code{Surv()} data using the estimate value of psi to give counterfactual untreated failure times.
\item ans: the object returned from \code{uniroot} used to solve the estimating equation
\item CI: a vector of the confidence interval around psi
\item call: the R call object
}
}
\description{
Main Function used for estmating causal parameters under the Rank Preserving Structural Failure Time Model
}
\details{
the formula object \code{ReCen(time, censor_time)~Instr(arm,rx)}, identifies particular meaning to the four
sets of arguments. \code{ReCen()} stands for ReCensoring. \code{Instr()} stands for Instrument. 
\itemize{
\item time: the observed failure or censoring time
\item censor_time: the time at which censoring would, or has occurred. This is provided for all observations
unlike standard Kaplan-Meier or Cox regression where it is only given for censored observations
\item arm: the randomised treatment arm. a factor with 2 levels, or numeric variable with values 0/1.
\item rx: the proportion of time on active treatment (arm=1 or the non-reference level of the factor)
}
Further adjustment terms can be added on the right hand side of the formula if desired, included \code{strata()}
or \code{cluster()} terms.
}
\examples{
library(rpsftm)
?immdef
fit <- rpsftm(ReCen(progyrs, censyrs)~Instr(imm,1-xoyrs/progyrs),immdef)
print(fit)
summary(fit)
plot(fit)

}
\author{
Simon Bond
}

