% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsftm.R
\name{rpsftm}
\alias{rpsftm}
\title{Rank Preserving Sturctural Failure Time Model}
\usage{
rpsftm(time, censor_time, rx, arm, data, adjustors = NULL, test = survdiff,
  lowphi = -10, hiphi = 10, alpha = 0.05, ...)
}
\arguments{
\item{time}{the observed failure or censoring time.}

\item{censor_time}{the theoretical censoring time, either observed or set after Time.}

\item{rx}{the proportion of Time spent on treatment}

\item{arm}{the randomisation that is independent of the recensored survival times}

\item{data}{a data.frame object containing the variables}

\item{adjustors}{a forumula object of covariates to adjust for: \code{~strata(A)+B*C}}

\item{test}{the survival regression function to calculate the z-statistic: survdiff, coxph, survreg}

\item{lowphi}{the lower limit of the range to search for the causal parameter}

\item{hiphi}{the upper limit of the range to search for the causal paramater}

\item{alpha}{the significance level used to calculate confidence intervals}

\item{target}{the value to subtract off from the z-statistic}

\item{\code{...}}{arguments to supply to the test function.}
}
\value{
a list of
\itemize{
\item phi the estimated parameter
\item fit the fitted survival object
\item Sstar the recensored \code{Surv()} data using the estimate parameter
\item ans the object return from \code{uniroot} 
\item CI a vector of the confidence interval around phi
\item call the R call object
}
}
\description{
Main Function used for estmating causal parameters under the Rank Preserving Structural Failure Time Model
}
\author{
Simon Bond
}

