% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsftm.R
\name{rpsftm}
\alias{rpsftm}
\title{Rank Preserving Structural Failure Time Model}
\usage{
rpsftm(time, censor_time, rx, arm, data, formula = ~1, test = survdiff,
  lowphi = -10, hiphi = 10, alpha = 0.05, treat_weight = 1,
  Recensor = TRUE, Autoswitch = TRUE, ...)
}
\arguments{
\item{time}{the observed failure or censoring time.}

\item{censor_time}{the theoretical censoring time, either observed or set after time.}

\item{rx}{the proportion of time spent on treatment}

\item{arm}{the randomised arm. Either a numerical indicator, with 0 as the placebo, or a factor with the lowest level as placebo.}

\item{data}{an optional data frame that contains variables}

\item{formula}{a forumula object of covariates to adjust for: \code{~strata(A)+B*C}}

\item{test}{the survival regression function to calculate the z-statistic: survdiff, coxph, survreg}

\item{lowphi}{the lower limit of the range to search for the causal parameter}

\item{hiphi}{the upper limit of the range to search for the causal paramater}

\item{alpha}{the significance level used to calculate confidence intervals}

\item{treat_weight}{an optional parameter that phi is multiplied by on an individual observation level to give
differing impact to treatment. The values are transformed by abs(.)/max(abs(.)) to ensure 1 is the largest weight.}

\item{Recensor}{a logical to use recensoring if set to TRUE. Default is TRUE.}

\item{Autoswitch}{a logical to autodetect cases of no switching. Default is TRUE}

\item{\code{...}}{arguments to supply to the test function.}
}
\value{
a list of
\itemize{
\item phi the estimated parameter
\item fit the fitted survival object
\item Sstar the recensored \code{Surv()} data using the estimate parameter
\item ans the object return from \code{uniroot} 
\item CI a vector of the confidence interval around phi
\item call the R call object
}
}
\description{
Main Function used for estmating causal parameters under the Rank Preserving Structural Failure Time Model
}
\author{
Simon Bond
}

