% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpsftm.R
\name{rpsftm}
\alias{rpsftm}
\title{Rank Preserving Structural Failure Time Model}
\usage{
rpsftm(T, A, R, C, Y, id, strata = NULL, psi = NULL, weights = NULL)
}
\arguments{
\item{T}{Interval specific contribution time. Specifically, it is the amount of time contributed by the patient for the observation of A recorded.}

\item{A}{Indicator of treatment observed, allowed to vary over time. For subjects who never cross over in treatment, this value should be the same as \code{R}.}

\item{R}{Indicator of treatment group initially randomized to. This is a binary variable that takes on value 1 if the subject was randomized to treatment and 0 if the subject was randomized to the control group. This can be either the same length as the length of the \code{T} and \code{A} vectors (and should not vary within each \code{id}) or the same length as there are unique \code{id} (and have the \code{id} as the names of each element in the vector).}

\item{C}{The observed censoring time. The model assumes that if a subject experiences a failure prior to observing the censoring time, then the censoring time would be the time from randomization until the end of study. It also assumes that censoring is independent of treatment assignment \code{R}. This can be either the same length as the length of the \code{T} and \code{A} vectors (and should not vary within each \code{id}) or the same length as there are unique \code{id} (and have the \code{id} as the names of each element in the vector).}

\item{Y}{Indicator of observed failure event. This is a binary variable that takes on value 1 if the event is observed and 0 otherwise. This can be either the same length as the length of the \code{T} and \code{A} vectors or the same length as there are unique \code{id}. If it is the same length as the \code{T} and \code{A} vectors, the function will check to see whether \code{Y} varies with \code{id}, i.e. whether there is more than 1 unique value of Y for each subject across all subjects. If so, then it is assumed that \code{Y} is a time-varying indicator of failure. An overall indicator will be computed from these time-varying values (with missing values removed beforehand). If \code{Y} is the same length as there are unique \code{id} values, then each element will require an \code{id} value name.}

\item{id}{Cluster identification variable. This is used to calculating the total amount of follow-up time per person. It should be of the same length as the \code{T} and \code{A} vectors.}

\item{strata}{Strata that each id is assigned to. This should be provided as a factor with the same length as the length of the \code{T} and \code{A} vectors and should not vary within each \code{id}.}

\item{psi}{Grid of potential parameter values to estimate over. By default, the grid is set to \code{seq(-3,3,.001)}.}

\item{weight}{Observational weights. This can be either the same length as the length of the \code{T} and \code{A} vectors or the same length as there are unique \code{id}. If \code{Y} is the same length as there are unique \code{id} values, then each element will require an \code{id} value name. If inserting data in longitudinal form, weights should correspond to weight for each observation (cf. cumulative weights over all time points). The cumulative product will then be taken over all time points.}
}
\value{
From the potential values of psi, the function will return an object of class \code{rpfstm}, which includes (1) the statistics for the counterfactual distribution, (2) the empirical values for the counterfactual distribution off treatment, and (3) the summarized data for time spent both on and off treatment. The data returned is ordered by \code{id}, with numeric IDs converted to character form. For the statistics, the following values are reported
\itemize{
	\item {log-rank statistic from the Kaplan Meier curve}
	\item {log-rank p-value from the Kaplan Meier curve}
	\item {Wilcoxon statistic from the Kaplan Meier curve}
	\item {Wilcoxon p-value from the Kaplan Meier curve}
	\item {Cox Wald statistic from the proportional hazards model}
	\item {Cox Wald p-value from the proportional hazards model}
	\item {Cox likelihood ratio test statistic from the proportional hazards model}
	\item {Cox likelihood ratio test p-value from the proportional hazards model}
	\item {Cox hazard ratio from the proportional hazards model}
}
}
\description{
\code{rpsftm} implements the rank preserving structural failure time model presented by Robins and Tsiatis (1991) and again by Korhonen et al (2012).
}
\details{
Data should be provided to the function in a longitudinal format, with the indicator of final outcome \code{Y} being constant across all observations from the same individual. \code{T} is the interval specific time contributed while on the specific treatment \code{TX}.
}
\author{
Linh Tran
}

