% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.rpsftm.R
\name{summary.rpsftm}
\alias{summary.rpsftm}
\title{summary Method}
\usage{
\method{summary}{rpsftm}(x)
}
\arguments{
\item{x}{an object returned from the \code{rpsftm()} function}
}
\value{
a summary of the fitted regression model
}
\description{
Function used to summarise the fitted model to an rpsftm object
}
\author{
Simon Bond
}

