% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untreated.R
\name{untreated}
\alias{untreated}
\title{Untreated Event Time}
\usage{
untreated(psi, response, treatment_matrix, rand_matrix, censor_time,
  autoswitch)
}
\arguments{
\item{psi}{the parameter that measures how more rapidly the lifetime is expended under treatment}

\item{response}{the response variable, a Surv() object}

\item{treatment_matrix}{a model.matrix object capturing the amount of active treatment(s) received}

\item{rand_matrix}{a model.matrix object catpuring the randomised treatment arms. Unused at presnet}

\item{censor_time}{the theoretical censoring time, either observed or set after time. Set to Inf to ignore recensoring.}

\item{autoswitch}{Currently has no effect ( formerly: a logical to autodetect cases of no switching. If TRUE, then if all observations in an arm
have perfect compliance then recensoring is not applied in that arm. If FALSE the recensoring is applied
regardless of perfect compliance.)}
}
\value{
A Surv() object
}
\description{
Calculates the counterfactual untreated event time, possibly using the recensoring method, assuming a given parameter value and returns a Surv() object
}
\author{
Simon Bond
}
\keyword{internal}
